Fix auth link between Document/Corpus/PyPLN objects
`PyPLN.corpora` was returning a list of `Copus` objects, but these objects did
not have `auth` set so you can't use them actually. In that case you needed to
pass `auth` in every call such as `add_document`.
Now we're using a `requests.Session` object to handl authentication so we don't
need to pass it in every action, only on class instantiation (it's valid for
`Corpus` and `Document` classes).